+++ lib/features/theory/domain/models/chord_identity.dart
@@ -125,6 +125,7 @@ class ChordIdentity {
 
 enum ChordQualityToken {
   major,
+  majorFlat5,
   minor,
   minorSharp5,
   diminished,